Move reset of bucket properties into bucket_properties_operator
